+++ dorklab/__init__.py
@@ -1,5 +1,5 @@
 """DorkLab - ricerca avanzata e protective dorking con interfaccia grafica."""
 
-__version__ = "1.5.0"
+__version__ = "1.6.0"
 APP_NAME = "DorkLab"
 APP_ID = "dorklab"